[#4596] Allow activity usage dialog to be skipped
Uses the various "Skip Dialog" keybindings to skip the activity
usage dialog, using the activity with the default configuration.
Does some modifications for the `ActivityChoiceDialog` to return
an event so the usage dialog can be skipped from there also.

The same keybindings will now also skip the activity choice dialog
rather than just being hardcoded to the Shift key. This does not
yet attempt to change the hint displayed in the UI of that dialog.

Closes #4596