[#2678] Fix "Reveal" button in secret blocks in multi-editor
Fixes the issues with reveal buttons for secret blocks for items
and actors that use multiple description editors. This required
overriding `_getSecretContent` and `_updateSecret` to look for
`data-target` in addition to `data-edit`. The locking code was
also modified to avoid disabling the secret button when the sheets
are in play mode.

There is a still a bug with the PC sheet and the item sheets with
only a single description (like classes), which don't retain the
secret state if the play/edit toggle is changed immediately after
changing the secret state. This appears to be because when it
submits `FormDataExtended` is fetching old data from the editor
instances on the sheet.

Closes #2678