Store scale value data in a sparse format
Changes how scale value data is stored from storing all keys for
each level to just storing the keys that were changed at a given
level.

```javascript
// Original
{
  1: { number: 1, faces: 6 },
  5: { number: 1, faces: 8 },
  11: { number: 1, faces: 10 }
}

// New
{
  1: { number: 1, faces: 6 },
  5: { faces: 8 },
  11: { faces: 10 }
}
```

This has the advantage of making it easier to change part of a
value. For example, if you had a scale value like Sneak Attack
that changes every other level and you wanted to modify it from
one die to two then it currently would require changing that one
in 10 places. With this change the number of dice would only need
to be updated at level 1.